test: relocate asset sidebar specs as git renames
Keep the unique-test union, but represent the two original sidebar
files as renames into tests/assets/ so CI does not treat relocated
coverage as newly added specs.
